requirements.txt
beautifulsoup4==4.9.0
discord.py==1.3.3
distro-info===0.18ubuntu0.18.04.1
youtube-dl==2020.3.24
